requirements.txt
quart
yt-dlp
aiofiles
aiohttp
bottom
tinytag
peewee
python-dateutil
mutagen
peewee
discord.py
